lap time fidelity switch to showing accuracy of 0.1s in onboard lap times also adds a way to test this : by hitting 'l' on keyboard also switches default screen display resolution to 1024x600

+++ lemon_pi/car/display_providers.py
@@ -40,7 +40,8 @@ class LapProvider():
     def get_lap_timer(self) -> int:
         pass
 
-    def get_last_lap_time(self) -> int:
+    # return seconds and millis for last lap
+    def get_last_lap_time(self) -> float:
         pass
